Revert "Merge branch 'feature_catapultapi' into feature_spatialMetricsModel"
This reverts commit 3270fbe496ecf436d2221501a7e2045a7f519e1d, reversing
changes made to 5cabaf0ba89ccf2181cee5b57a7530952f1db21e.
