--- docs/source/modules/io/catapult.rst
@@ -1,6 +1,0 @@
-=====================
-floodlight.io.catapult
-=====================
-
-.. automodule:: floodlight.io.catapult
-    :members:
