--- docs/source/modules/io/catapultApi.rst
@@ -1,6 +1,0 @@
-=====================
-floodlight.io.catapultApi
-=====================
-
-.. automodule:: floodlight.io.catapultApi
-    :members:
