+++ docs/source/modules/io/io.rst
@@ -21,21 +21,6 @@ Collection of file parsing functionalities for different data providers as well 
    utils
 
 
-.. rubric:: catapult
-   
-.. currentmodule:: floodlight.io.catapult
-.. autosummary::
-   :nosignatures:
-
-   dump_list_of_dicts
-   get_activity_players_info
-   get_meta_data
-   get_players_sensor_data_dict_list
-   load_list_of_dicts
-   read_position_data_from_activity
-   read_position_data_from_dict_list
-
-   
 .. rubric:: Datasets
 
 .. currentmodule:: floodlight.io.datasets
